close ticket

+++ modules/Abd/Ticket/src/Resources/Views/index.blade.php
@@ -37,11 +37,12 @@
                             <td>{{$ticket->updated_at}}</td>
                             <td>{{$ticket->status}}</td>
                             <td>
-                                <a href=""
-                                   onclick="deleteItem(event, '{{ route('tickets.destroy', $ticket->id)}}');"
-                                   class="item-delete mlg-15" title="حذف"></a>
-                                <a href="{{ route('tickets.edit', $ticket->id) }}" class="item-edit mlg-15"
-                                   title="ویرایش"></a>
+                                <a href="{{route("tickets.close", $ticket->id)}}">بستن تیکت</a>
+                                @can(\Abd\RolePermissions\Models\Permission::PERMISSION_MANAGE_TICKETS)
+                                    <a href=""
+                                       onclick="deleteItem(event, '{{ route('tickets.destroy', $ticket->id)}}');"
+                                       class="item-delete mlg-15" title="حذف"></a>
+                                @endcan
                             </td>
                         </tr>
                     @endforeach
